move files, update tests in neighborhood
