+++ python/pythonlab/neighborhood/neighborhood/__init__.py
@@ -1,4 +1,7 @@
 from .painter import Painter as Painter
+from .neighborhood_log import NeighborhoodLog as NeighborhoodLog
+from .painter_log import PainterLog as PainterLog
+from .painter_event import PainterEvent as PainterEvent
 # We export World so we can clear the world in pythonlab_setup and
 # handle setting the context type in unittest_runner
 from .support.world import World as World
