+++ python/pythonlab/neighborhood/neighborhood/painter_log.py
@@ -1,5 +1,5 @@
-from .position import Position
-from .neighborhood_signal_message import NeighborhoodSignalMessage
+from .support.position import Position
+from .support.neighborhood_signal_message import NeighborhoodSignalMessage
 from .painter_event import PainterEvent
 
 class PainterLog:
