+++ python/pythonlab/neighborhood/tests/support/test_neighborhood_tracker.py
@@ -11,11 +11,15 @@ def test_initialize_neighborhood_tracker():
     neighborhood_world.set_grid_from_string(SAMPLE_MAZE)
     neighborhood_world.set_context_type(NeighborhoodContextType.VALIDATE)
     neighborhood_tracker = NeighborhoodTracker(neighborhood_world)
-    neighborhood_log = neighborhood_tracker.get_neighborhood_log()
-    assert neighborhood_log == 'neighborhood_log'
-    initialize_painter_signal_message = NeighborhoodSignalMessage(SignalMessageType.NEIGHBORHOOD, NeighborhoodSignalKey.INITIALIZE_PAINTER, {"id": 1})
+    initialize_painter_signal_message = NeighborhoodSignalMessage(SignalMessageType.NEIGHBORHOOD, NeighborhoodSignalKey.INITIALIZE_PAINTER, {"id": "painter-1", "x": 0, "y": 0, "paint": 2, "direction": "east"})
     assert neighborhood_tracker.is_initialized is False
     neighborhood_tracker.track_signal(initialize_painter_signal_message)
+    neighborhood_log = neighborhood_tracker.get_neighborhood_log()   
+    assert len(neighborhood_log.painter_logs) == 1
+    grid_size = 2
+    expected_neighborhood_state = [[None for _ in range(grid_size)] for _ in range(grid_size)]
+    assert neighborhood_log.final_output_matches(expected_neighborhood_state) is True
+    assert len(neighborhood_log.painter_logs) == 1
     assert neighborhood_tracker.is_initialized is True
     neighborhood_tracker.reset()
     assert neighborhood_tracker.is_initialized is False
