+++ python/pythonlab/neighborhood/tests/support/test_painter_tracker.py
@@ -6,7 +6,7 @@ from neighborhood.support.position import Position
 
 def test_initialize_painter_tracker():
     painter_tracker = PainterTracker("painter-1", Position(0,0,'East'), 2)
-    assert painter_tracker.get_painter_log() == 'painter_log'
+    assert painter_tracker.get_painter_log().painter_id == 'painter-1'
     move_signal_message = NeighborhoodSignalMessage(SignalMessageType.NEIGHBORHOOD, NeighborhoodSignalKey.MOVE, {"direction": "East"})
     assert painter_tracker.current_position.x == 0
     assert painter_tracker.current_position.y == 0
