+++ python/pythonlab/unittest_runner/unittest_runner/validation_protocol.py
@@ -27,10 +27,8 @@ class ValidationProtocol(object):
     self._invoke_main(file_path)
 
     World().set_context_type(NeighborhoodContextType.RUN)
-    neighborhood_log = self.neighborhood_tracker.get_neighborhood_log()
+    neighborhood_log = NeighborhoodTracker._instance.get_neighborhood_log()
     self.neighborhood_tracker.reset()
-    # Currently returns None.
-    # TODO: Actually return neighborhood log
     return neighborhood_log
